% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodTuringProportions.R
\name{goodTuringProportions}
\alias{goodTuringProportions}
\title{Good Turing Proportions}
\usage{
goodTuringProportions(counts)
}
\arguments{
\item{counts}{}
}
\value{
matrix of GT proportions
}
\description{
Calculatest the Good Truing proportions for a matrix of expression values
}
\examples{
blabla

}
