% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTestimate.R, R/zzz.R
\docType{package}
\name{GTestimate}
\alias{GTestimate}
\title{GTestimate}
\usage{
GTestimate(object, ...)
}
\arguments{
\item{x:}{input expression vector}

\item{conf:}{confidence value for model switch}
}
\value{
a named list containing
P0: the proportion of missing mass
proportion: the estimated proportions for the carious count levels
count: the count levels
n: the number of occurences for each count level
n0: the number of 0zeroes in the input
}
\description{
Simple Good-Turing algorithm for frequency estimation
as described by Gale and Sampson (1995)
Sampson's C code translated to C++ and optimized by Aaron Lun
Has been tested against Sampson's C code from
http://www.grsampson.net/RGoodTur.html
and gives identical results.
Gordon Smyth and Aaron Lun
9 Nov 2010.  Last modified 7 Sep 2012.

Description of your package
}
\examples{
library(Seurat)
data('pbmc_small')
GTestimate(pbmc_small)
}
\author{
Martin Fahrenberger
}
