% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTestimate.R, R/zzz.R
\docType{package}
\name{GTestimate}
\alias{GTestimate}
\title{GTestimate}
\usage{
GTestimate(object, ...)
}
\description{
Wrapper functions for Good Turing estimation of scRNA-seq relative gene expression estimation.

Description of your package
}
\examples{
library(Seurat)
data('pbmc_small')
GTestimate(pbmc_small)
}
\author{
Martin Fahrenberger
}
